Reduce number of created layers
Every time "RUN" "ADD" etc are used another layer is created. So as a suggestion, use multi-line one-liners to reduce the number of layers. RUN of yum clean in particular creates a layer that masks over the junk in the lower layers. So it looks like yum caches are cleared but they arent.

The list of ADD's needs more thought, but the RUN's all mixed in together could be moved upwards into the single long RUN command