tests: row_cache_test: Fix test failure
Broken after 0ac2c388b6d3e6d36747cbe8088ed794bfa13a6d, which assigns
empty reader to _delegate on hitting wide partition limit. The test
assumed that the original _delegate will be invoked when the
single-partition reader is asked for the next partition, which is no
longer the case.

Message-Id: <20170912172739.6851-1-tgrabiec@scylladb.com>